Vistas de Almuerzo

+++ SmartFoodV2.0/SmartFood/src/smartfood/models/AlmuerzoDecorator.java
@@ -14,6 +14,6 @@ public abstract class AlmuerzoDecorator extends Almuerzo {
     protected Almuerzo almuerzo;
     
     @Override
-    public abstract float getCosto(); 
+    public abstract double getCosto(); 
     
 }
